prep run 4: label smoothing, float16 encoder features, run script
- Add label_smoothing support to segmentation cross-entropy loss
- Save encoder features as float16 (halves disk: ~43GB vs ~86GB for humanrig)
- Add segmentation_a100_run4.yaml (resume from run 1, lower LR, stronger aug)
- Add weights_a100_run4.yaml (same as lean, for split loader fix)
- Add training/run_fourth.sh (full run 4 pipeline)

+++ training/data/precompute_encoder_features.py
@@ -252,8 +252,8 @@ def main() -> None:
                 feature_map, vertex_positions, image_size[0], image_size[1]
             )
 
-            # Save
-            np.save(out_path, sampled)
+            # Save as float16 to halve disk usage (~3.5MB vs ~7.5MB per file)
+            np.save(out_path, sampled.astype(np.float16))
             processed += 1
 
             if processed % 500 == 0:
